Separate out code from run.h
   * bench mark result data from run.h goes in bench_results.h
   * output code from run.h goes in output.h, which can use bench results
   * code that generates data for benchmarking goes into data_sources.h from run.h.
